Allow passing multiple YAML files to dev_appserver.py
In a scenario where you have multiple modules (API backend + SPA frontend) with a dispatch.yaml, with the current setup it is not possible to pass multiple yaml paths to the dev_appserver.py call. Additionally, there are options that are not prefixed with '--' but rather a single '-' as is the case with the '-A' parameter for the dev_appserver.py (applicable for Flexible runtime environments). This hack will work so long as the caller prioritizes the YAML files before any additional options in the commands option on the call. Open to more elegant solutions.